new mock_runtime, and tidy up tests. Started on CoW+S3 migration test
